Separating model loading from importing
